+++ src/Cache/Options/IOptions.cs
@@ -1,3 +1,5 @@
+using Newtonsoft.Json;
+
 namespace GerwimFeiken.Cache.Options;
 
 public interface IOptions
@@ -6,4 +8,9 @@ public interface IOptions
     /// Defaults to 86400 (1 day).
     /// </summary>
     public int DefaultExpirationTtl { get; set; }
+
+    /// <summary>
+    /// (De)serialization settings. If set, this will overwrite the default settings.
+    /// </summary>
+    JsonSerializerSettings? JsonSerializerSettings { get; set; }
 }